--- charts/output.csv
@@ -1,7 +1,0 @@
-Filesize;Time;;
-100;1;;
-1000;5;;
-2000;10;;
-3000;70;;
-4000;450;;
-5000;520;;
