--- charts/reference.csv
@@ -1,7 +1,0 @@
-Filesize;Time;;
-100;1;;
-1000;2;;
-2000;5;;
-3000;40;;
-4000;400;;
-5000;500;;
