+++ charts/requriments.txt
@@ -1,1 +1,1 @@
-mathpoltlib
+matplotlib
