--- charts/table1.tex
@@ -1,21 +1,0 @@
-\documentclass{article}
-\usepackage[utf8]{inputenc}
-\usepackage{graphicx}\usepackage{subcaption}
-\usepackage{rotating}
-\begin{document}\begin{table}
-\centering
-\begin{tabular}{lll}
- \hline
-Filesize & Time our& Time FPGA\\ \hline 
-100&1&1\\ 
-1000&5&2\\ 
-2000&10&5\\ 
-3000&70&40\\ 
-4000&450&400\\ 
-5000&520&500\\ 
-\hline
-\end{tabular}
-\caption{FIX ME!!!}
-\label{tbl:oprv}
-\end{table}
-\end{document}
